test(backend): Refactor ArchUnit package further
