+++ backend/src/test/java/dev/codesoapbox/backity/archunit/production/rules/UseCaseRules.java
@@ -1,10 +1,11 @@
-package dev.codesoapbox.backity.archunit.rules;
+package dev.codesoapbox.backity.archunit.production.rules;
 
 import com.tngtech.archunit.junit.ArchTest;
 import com.tngtech.archunit.lang.ArchRule;
 
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.methods;
 
+/// Rules specifically about use cases, unrelated to anything else.
 @SuppressWarnings("unused")
 public class UseCaseRules {
 
@@ -19,7 +20,8 @@ public class UseCaseRules {
                     the use case should be adequately described by its class name.
                     
                     Context:
-                    Use cases classes should follow the `<Action>UseCase` naming convention and contain only one method.
+                    Use cases classes should follow the `<Action>UseCase` naming convention \
+                    and contain only one method. \
                     Therefore, providing a method name other than `execute` would be superfluous and may lead to \
                     class and methods diverging over time.
