+++ backend/src/test/java/dev/codesoapbox/backity/archunit/test/ArchUnitTestCodeTest.java
@@ -1,4 +1,4 @@
-package dev.codesoapbox.backity.archunit;
+package dev.codesoapbox.backity.archunit.test;
 
 import com.tngtech.archunit.core.importer.ImportOption;
 import com.tngtech.archunit.junit.AnalyzeClasses;
@@ -6,7 +6,8 @@ import com.tngtech.archunit.junit.ArchTest;
 import com.tngtech.archunit.junit.ArchTests;
 import com.tngtech.archunit.junit.CacheMode;
 import dev.codesoapbox.backity.BackityApplication;
-import dev.codesoapbox.backity.archunit.rules.TestRules;
+import dev.codesoapbox.backity.archunit.test.rules.LowLevelTestRules;
+import dev.codesoapbox.backity.archunit.test.rules.TestReadabilityRules;
 
 @SuppressWarnings("unused")
 @AnalyzeClasses(packagesOf = BackityApplication.class, importOptions = ImportOption.OnlyIncludeTests.class,
@@ -15,5 +16,8 @@ import dev.codesoapbox.backity.archunit.rules.TestRules;
 public class ArchUnitTestCodeTest {
 
     @ArchTest
-    static final ArchTests TEST_RULES = ArchTests.in(TestRules.class);
+    static final ArchTests LOW_LEVEL_RULES = ArchTests.in(LowLevelTestRules.class);
+
+    @ArchTest
+    static final ArchTests READABILITY = ArchTests.in(TestReadabilityRules.class);
 }
